HSEARCH-2886 Move ProgressiveCharBufferWriter to its own file
